backend/requirements.txt
Django==5.1.5
gunicorn==23.0.0
python-slugify==8.0.4
django-mptt==0.16.0
